Moved namespaces for FilteredStrings
